Refactor main.go
- Tests remain in the main_test.go since they mostly require access to
the mock server.
